enhance pull request 2618
