--- dubbo-serialization/dubbo-serialization-protobuf/src/main/resources/META-INF.dubbo.internal/org.apache.dubbo.common.serialize.Serialization
@@ -1,1 +1,0 @@
-protobuf=org.apache.dubbo.common.serialize.protobuf.ProtobufSerialization
