+++ dubbo-serialization/dubbo-serialization-protostuff/src/main/java/org/apache/dubbo/common/serialize/protostuff/utils/WrapperUtils.java
@@ -15,12 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.common.serialize.protobuf.utils;
+package org.apache.dubbo.common.serialize.protostuff.utils;
 
-import org.apache.dubbo.common.serialize.protobuf.Wrapper;
+import org.apache.dubbo.common.serialize.protostuff.Wrapper;
 
 import java.math.BigDecimal;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.BitSet;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Hashtable;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.SortedMap;
+import java.util.TreeMap;
+import java.util.TreeSet;
+import java.util.Vector;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class WrapperUtils {
