+++ dubbo-serialization/dubbo-serialization-protostuff/src/main/resources/META-INF.dubbo.internal/org.apache.dubbo.common.serialize.Serialization
@@ -1,0 +1,1 @@
+protostuff=org.apache.dubbo.common.serialize.protostuff.ProtostuffSerialization
