+++ dubbo-serialization/dubbo-serialization-test/src/test/java/org/apache/dubbo/common/serialize/protostuff/ProtostuffSerializationTest.java
@@ -15,19 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.common.serialize.protobuf;
+package org.apache.dubbo.common.serialize.protostuff;
 
-/**
- * Protostuff can only serialize/deserialize POJOs, for those it can't deal with, use this Wrapper.
- */
-public class Wrapper<T> {
-    private T data;
+import org.apache.dubbo.common.serialize.base.AbstractSerializationTest;
 
-    Wrapper(T data) {
-        this.data = data;
-    }
-
-    Object getData() {
-        return data;
+public class ProtostuffSerializationTest extends AbstractSerializationTest {
+    {
+        serialization = new ProtostuffSerialization();
     }
 }
